lexer/lexer_test.go
package lexer

import (
	"testing"

	"github.com/technoboom/compiler/token"
)

func TestNewToken(t *testing.T) {
	input := "{=+(),;-*/}"
	tests := []struct {
		expectedType    token.Type
		expectedLiteral string
	}{
		{token.LBRACKET, "{"},
		{token.ASSIGN, "="},
		{token.PLUS, "+"},
		{token.LPAREN, "("},
		{token.RPAREN, ")"},
		{token.COMMA, ","},
		{token.SEMICOLON, ";"},
		{token.MINUS, "-"},
		{token.MULTIPLY, "*"},
		{token.DIVIDE, "/"},
		{token.RBRACKET, "}"},
	}

	l := New(input)

	for i, tt := range tests {
		tok := l.NextToken()

		if tok.Type != tt.expectedType {
			t.Fatalf("tests[%d] - token type wrong. expected=%q, got=%q",
				i, tt.expectedType, tok.Type)
		}

		if tok.Literal != tt.expectedLiteral {
			t.Fatalf("tests[%d] - token literal wrong. expected=%q, got=%q",
				i, tt.expectedLiteral, tok.Literal)
		}
	}

	// check EOF
	eofToken := l.NextToken()
	if eofToken.Type != token.EOF {
		t.Fatalf("test for EOF failed. expected=EOF, got=%q", eofToken.Type)
	}
}
